backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
xgboost>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
pyjwt>=2.8.0
passlib[bcrypt]>=1.7.4
pydantic>=2.0.0
python-multipart>=0.0.6
